pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "panier-cli"
version = "0.1.0"
description = "CLI de planification repas et courses optimisées multi-drive avec préférences, allergies et budget."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Julien" }]
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.5.0",
]

[project.scripts]
panier = "panier.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/panier"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
"src/panier/cart.py" = ["E501"]
